requirement.txt
fastapi
sqlalchemy
uvicorn
bcrypt
passlib
